# Add isolate feature