--- a/tools/isolate/index.js
+++ b/tools/isolate/index.js
@@ -1,10 +1,11 @@
 const util = require('util');
 const path = require('path');
-const glob = util.promisify(require('node-glob'));
+const glob = util.promisify(require('glob'));
 const mv = require('move-file');
 
 async function run() {
-  const matches = await glob('src/site/content/**/*.md');
+  const ignore = process.argv.slice(2);
+  const matches = await glob('src/site/content/**/*.md', {ignore});
   matches.forEach(async (oldPath) => {
     let newPath = oldPath.split(path.sep);
     newPath.splice(2, 1, '_exile');