This adds a (very experimental) feature to improve the quality of life for folks who are developing or authoring content on web.dev.

`npm run isolate` can be used to move all files out of `src/site/content` and into an `_exile` directory which will be ignored for builds. `isolate` takes an optional glob (or space separated list of globs) of files that should be preserved.

`npm run integrate` will restore all of the files in `_exile` to their previous location. It will also delete the `_exile` directory.

Example usage:

```bash
# Move everything except the style-focus directory to _exile
npm run isolate -- src/site/content/en/accessible/style-focus/**
```

(Do some work on style-focus/index.md)

```bash
# Restore everything from _exile back to src/site/content
npm run integrate
```

## Benchmark

Without isolate:
- Initial Eleventy build: **10.48 seconds**
- Eleventy rebuild: **8.27 seconds**

With isolate:
- Initial Eleventy build: **1.06 seconds**
- Eleventy rebuild: **0.53 seconds**

## Changes

Changes proposed in this pull request:

- Adds `isolate` and `integrate` npm scripts
- Adds git commit hook to prevent committing while the `_exile` directory is active
- Prevents authors collection from throwing if there are authors without posts