@robdodson 2020-09-22
Percy failure can be ignored. It's flaking on an image (related to #3804).