@nhoizey 2020-09-23
Wouldn't it be enough to restrict the number of items in every collection when the environment is "dev"?