@robdodson 2020-09-23
> Wouldn't it be enough to restrict the number of items in every collection when the environment is "dev"?

Maybe, though I think that would require editing the code for all of our collections to use a common filter and we'd need to ensure we use it for new collections as well. It's not a bad idea though and may be safer than physically moving the files (though that seemed to work surprisingly well). I'd also need a way to specify to the filter that it should allow certain posts to remain in the collection(s) (basically whichever post or directory of posts I'm working on).