@nhoizey 2020-09-23
I agree it might not be easy, and you probably don't want to remove items from a navigation collection, for example… 😅